requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
astunparse==1.6.3
blis==0.9.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.0
gast==0.5.4
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
path==16.7.1
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.7.0
pyparsing==3.1.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post9
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thinc==8.2.0
threadpoolctl==3.2.0
tokenizers==0.14.0
tqdm==4.66.1
transformers==4.33.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.7
wrapt==1.15.0
openai
pandas
